Rough pass of refactoring
Replace all instances of AddressBook with TodoList (case-sensitive).
Be aware that while some files have been renamed to ensure that the
build passes, there exists files and variables that have not been caught
in this sweeping change. This is intentional. It demonstrates that
automated refactoring and autocomplete is not infallible.

Fix checkstyle violations as imports may be ordered wrongly after the
renaming.